% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powrballad-package.r
\docType{data}
\name{ballads}
\alias{ballads}
\title{A list of Power Ballads that can be used to randomly select from}
\format{An object of class \code{data.frame} with 38 rows and 1 columns.}
\usage{
ballads
}
\description{
A list of Power Ballads that can be used to randomly select from
}
\keyword{datasets}
