% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\name{select_powerballad}
\alias{select_powerballad}
\title{Select a powerballad, randomly}
\usage{
select_powerballad()
}
\value{
a character vector of length 1 containing a song and artist
}
\description{
Use this function to randomly select a power ballad
from the list.
}
\examples{
select_powerballad()
}
